requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2022.9.24
cffi==1.17.1
channels==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2.17
django-channels-notifications==1.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-dbbackup==4.0.2
django-debug-toolbar==3.7.0
django-import-export==2.9.0
django-material==1.12.0
django-material-admin==1.8.6
django-mathfilters==1.0.0
django-silk==5.4.3
django-user-messages==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
excel2json-3==0.1.6
gprof2dot==2025.4.14
httpie==3.2.1
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
jsonfield==3.1.0
lml==0.1.0
markdown-it-py==2.1.0
MarkupPy==1.14
material==0.1
mdurl==0.1.2
multidict==6.0.4
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==24.3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.3.1
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2022.4
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
zope.interface==7.2
